pyproject.toml
[tool.poetry]
name = "rag-chatbot"
version = "0.1.0"
description = ""
authors = ["Võ Đình Đạt <vodinhdat03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gradio = "<5.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
llama-index-embeddings-huggingface = "^0.1.4"
llama-index-vector-stores-chroma = "^0.1.6"
llama-index-llms-ollama = "^0.1.2"
llama-index-readers-file = "^0.1.11"
llama-index = "^0.10.22"
transformers = "^4.39.1"
torch = "*"
pymupdf = "^1.24.3"
tqdm = "^4.66.4"
requests = "^2.32.3"
pandas = "^2.2.3"
sentence-transformers = "^3.2.0"



[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
deptry = "^0.20.0"
pytest = "^8.3.3"

[tool.deptry.per_rule_ignores]
DEP002 = ["sentence-transformers"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
